Rename a file on the client side so client and server have modules of the same name
